pyproject.toml
[tool.poetry]
name = "pico-display"
version = "0.1.0"
description = "Scripts to display info on waveshare 5.65 display"
authors = ["Arjun Naik"]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.1"
Pillow = "^9.2.0"
icalendar = "^4.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
